add sd-notify to signal readyness
Provide a way to signal nsncd has successfully started up, and is ready
to accept connections.

A more reliable way would be to only signal this once at least one
worker has started up, but this is good enough (tm) for now.

This can be used in a systemd.service file with Type=notify.