Extend the result vector by a slice rather than pushing bytes
This makes the code a bit cleaner and gives the stdlib / compiler more
information about what we actually want to achieve. The previous
assembly looked rather bulky when recreating it on godbolt.

Running some hacky criterion benchmarks (after marking handlers public,
changing the entire thing to be a library with a binary rather than just
a binary, ..) the performance gains are roughly 20% on serializing host
results.