implement host lookup support
This adds support for the GETAI, GETHOSTBYADDR, GETHOSTBYADDRv6,
GETHOSTBYNAME, GETHOSTBYNAMEv6 request types.

For the more complex GETAI lookup, we use the dns_lookup crate.

In previous iterations of this change, we also used the same underlying
getaddrinfo call to respond to the  gethostbyname/gethostbyaddr
operations.

Even though gethostbyname/gethostbyaddr officially are deprecated,
there's a lot of tools still using it, and relying on them behaving
differently.

So it's important to still implement it, with exactly the same
behaviour, to prevent some tools from breaking (like `hostname --fqdn`,
see https://github.com/nix-community/nsncd/issues/4).

We FFI the right Glibc gethostbyname_2r/gethostbyaddr_2r (now
deprecated) functions and use it to back the GETHOSTBYNAME,
GETHOSTBYNAME6, GETHOSTBYADDR, and GETHOSTBYADDR6 Nscd interfaces.

Took me three try to get this right. This is actually the third
full rewrite.

The Nscd behaviour for these two legacy functions is *really*
confusing. We're supposed to ignore the herrno (herrno != errno!!) and
set it to 0 if gethostbyaddr/name returns a non-null hostent. If we
end up with a null hostent, we return the herrno together with a dummy
hostent header.

I tried to keep things as safe as possible by extracting the glibc
hostent to a proper rust structure. This structure mirrors the libc
hostent in a Rust idiomatic way. We should probably try to upstream
the FFI part of this commit to the Nix crate at some point.

Fixes https://github.com/nix-community/nsncd/issues/4

Co-Authored-By: Florian Klink <flokli@flokli.de>